rename locust
